+++ .vscode/tasks.json
@@ -4,7 +4,7 @@
         {
             "label": "build",
             "type": "shell",
-            "command": "gradlew buildFatJar -x test",
+            "command": "./gradlew buildFatJar -x test",
             "problemMatcher": [
                 "$gradle"
             ],
@@ -15,7 +15,7 @@
         {
             "label": "run",
             "type": "shell",
-            "command": "gradlew run -Pdevelopment=true",
+            "command": "./gradlew run -Pdevelopment=true",
             "problemMatcher": [
                 "$gradle"
             ],
@@ -26,7 +26,7 @@
         {
             "label": "run continuous",
             "type": "shell",
-            "command": "gradlew run --continuous -Pdevelopment=true -Pkotlin.incremental.js.ir=true",
+            "command": "./gradlew run --continuous -Pdevelopment=true -Pkotlin.incremental.js.ir=true",
             "problemMatcher": [
                 "$gradle"
             ],
@@ -38,7 +38,7 @@
         {
             "label": "test",
             "type": "shell",
-            "command": "gradlew test",
+            "command": "./gradlew test",
             "problemMatcher": [
                 "$gradle"
             ],
