+++ src/jvmMain/kotlin/net/pelata/features/pace/controller/PaceRequest.kt
@@ -1,4 +1,4 @@
-package net.pelata.features.pace.data
+package net.pelata.features.pace.controller
 
 import io.konform.validation.*
 import io.konform.validation.jsonschema.*
