--- src/jvmMain/kotlin/net/pelata/features/pace/data/Result.kt
@@ -1,13 +1,0 @@
-package net.pelata.features.pace.data
-
-import net.pelata.features.pace.data.SplitTime
-import net.pelata.units.Distance
-
-data class Result(
-    val unit: Distance,
-    val average: SplitTime,
-    val averageSpeed: Double,
-    val distances: List<Double>,
-    val splits: List<List<SplitTime>>,
-    val isFast: Boolean
-)
