+++ src/jvmMain/kotlin/net/pelata/features/pace/entity/SplitDistance.kt
@@ -1,0 +1,3 @@
+package net.pelata.features.pace.entity
+
+data class SplitDistanceList(val splitDistances : List<Double>)
