+++ src/jvmMain/kotlin/net/pelata/features/pace/entity/SplitTimeList.kt
@@ -1,0 +1,3 @@
+package net.pelata.features.pace.entity
+
+data class SplitTimeList(val splitTimes : List<SplitTime>)
