+++ src/jvmMain/kotlin/net/pelata/features/pace/entity/SplitTimeTable.kt
@@ -1,0 +1,5 @@
+package net.pelata.features.pace.entity
+
+import net.pelata.features.pace.entity.SplitTimeList
+
+data class SplitTimeTable(val splitTimeTable : List<SplitTimeList>)
